Creating packages for each k8s object.
